ipmi_<name> facts from first found LAN channel
I find some of my servers have ipmi running on channel #2 only (plugged into different physical port maybe - who knows).
so the impi_ipaddress etc facts (which come from channel #1) are not set. as channel #1 is not configured.

instead of always taking channel #1 for these, could we take 'the first found LAN channel' instead? (asside: this is the behaviour when running 'ipmitool lan print' without picking a channel).

have put in a suggested change for this. think it would work to only set each of the ipmi_<name> facts if not set so far.

cheers,