Add ability to control managing access
In certain IPMI implementations, you are not allowed to make any
access changes to the factory shipped user, for example, the
ADMIN user in Supermicro IPMI[1].

This results in the `setaccess` giving an error when trying to
manage the passwords for the default users.  This patch allows
a user to avoid running the access commands but defaults to
true to maintain backwards compatiblity.

[1]: http://www.supermicro.com/support/faqs/faq.cfm?faq=19375