ipmitool_fru: Add new fact to show to information from ipmtool fru
This CR adds a new structured fact called ipmitool with two keys fru and
mc_info.  fru is new data from ipmitool fru print.  mc_info is the
current ipmi_mc_info  fact.  I have updated the ipmi_mc_info fact so
that it uses this new structured fact.

The new fact will look like the following:

```
{
  fru => {
    board_mfg_date => "Tue Mar  3 21:43:00 2015",
    board_mfg => "DELL",
    board_product => "PowerEdge R220",
    board_serial => "*********",
    board_part_number => "0DRXF5A04",
    product_manufacturer => "DELL",
    product_name => "Test",
    product_extra => "*********"
  },
  mc_info => {
    IPMI_Puppet_Service_Recommend => "running",
    Device ID => "32",
    Device Revision => "1",
    Firmware Revision => "2.65",
    IPMI Version => "2.0",
    Manufacturer ID => "674",
    Manufacturer Name => "DELL Inc",
    Product ID => "256 (0x0100)",
    Product Name => "Unknown (0x100)",
    Device Available => "yes",
    Provides Device SDRs => "yes"
  }
}
```

We are able to use commands like the following:
set the board_product with:
* sudo ipmitool fru edit 0 field b 1

And the product_name with:
* sudo ipmitool fru edit 0 field p 1